Fix #20 crashed with SIGSEGV at rmd_get_frame.c:481
Fix verified by simulating `XFixesGetCursorImage(pdata->dpy)` returning
`NULL` sometimes by adding this code below it

```
                if (counter++ % 5 != 0)
                    xcim = NULL;
```
where `counter` is just

```
static int counter = 0;
```

and then making a recording with that code. Both regularly:

```
% recordmydesktop
```

and with `--follow-mouse`:

```
recordmydesktop --follow-mouse --width=256 --height=256
```

The result of this fix in the video recording is that the cursor
disappears (and `--follow-mouse` position is not updated) for the
frame(s) where NULL is returned.